requirements.txt
# maybe some of the following arent needed
# the 'done' are necessary
asgiref==3.2.10
certifi==2022.12.7
chardet==3.0.4
click==8.1.3
Django==3.1.14
django-crispy-forms==1.9.0
django-static-jquery==2.1.4
Flask==2.3.2 # done
Flask-Cors==3.0.9 # done
idna==2.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
prompt-toolkit==1.0.14
pycryptodome==3.9.7 # done
Pygments==2.7.4
PyInquirer==1.0.3
pytz==2019.3
regex==2020.2.20
requests==2.25.1 # done
six==1.14.0
sqlparse==0.4.4
texttable==1.6.2
urllib3==1.26.5
wcwidth==0.1.8
Werkzeug==2.3.3
wincertstore==0.2
numpy==1.19.5
